dashboard/server/requirements.txt
aniso8601==9.0.1
antiorm==1.2.1
Babel==2.9.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
db==0.1.1
db-sqlite3==0.0.1
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
Flask-User==1.0.2.2
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.0.4
importlib-metadata==3.7.3
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
passlib==1.7.4
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.4.2
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
speaklater==1.3
SQLAlchemy==1.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1
